Current state of development

+++ database_init/setup.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS playback_enq.discord_users (
-    id int NOT NULL,
+    id bigint NOT NULL,
     tag varchar(80),
     access_token varchar(255),
     refresh_token varchar(255),
@@ -8,8 +8,8 @@ CREATE TABLE IF NOT EXISTS playback_enq.discord_users (
 );
 
 CREATE TABLE IF NOT EXISTS playback_enq.party_channels (
-    id int NOT NULL,
-    owner_id int NOT NULL,
+    id bigint NOT NULL,
+    owner_id bigint NOT NULL,
     autodelete_date DATETIME,
     PRIMARY KEY (id),
     CONSTRAINT fk_channel_owner FOREIGN KEY (owner_id) REFERENCES discord_users(id)
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS playback_enq.party_channels (
 
 CREATE TABLE IF NOT EXISTS playback_enq.code_requests (
     request_id int NOT NULL AUTO_INCREMENT,
-    discord_user int NOT NULL,
+    discord_user bigint NOT NULL,
     request_state varchar(255) NOT NULL,
     code varchar(255),
     PRIMARY KEY (request_id),
